Convert GUI/App to two-part. Persons table change
Kept GUI related code in dbpl_search_and_draw_GUI_, tranferred rest of
the code in new file dbpl_search_and_draw.

Also, changed id in Persons table back to integer. URL is retained as
the unique external key. Dropped Name and Surname columns from Persons
table.
